Get_All_Movies_Works

+++ create_data.py
@@ -20,7 +20,7 @@ class Movie(db.Model):
     description = db.Column(db.String(255))
     trailer = db.Column(db.String(255))
     year = db.Column(db.Integer)
-    rating = db.Column(db.Integer)
+    rating = db.Column(db.Float)
     genre_id = db.Column(db.Integer, db.ForeignKey("genre.id"))
     genre = db.relationship("Genre")
     director_id = db.Column(db.Integer, db.ForeignKey("director.id"))
@@ -254,6 +254,7 @@ for movie in data["movies"]:
         rating=movie["rating"],
         genre_id=movie["genre_id"],
         director_id=movie["director_id"],
+
     )
     with db.session.begin():
         db.session.add(m)
